Rewrite short form functions
Was having issues including this template in a cdk CfnInclude due to the
short form functions (see https://github.com/aws/aws-cdk/issues/6537).
Changed to use long form. I would have liked to fix this on the cdk side,
but the work they're doing on this is in an experimental module, and
these changes shouldn't break anything in the source template.